feat: queryDsl로 검색 where문

+++ umc_9th_paulo/src/main/java/com/example/umc_9th_paulo/domain/review/dto/ReviewRequestDto.java
@@ -8,7 +8,7 @@ public class ReviewRequestDto {
     @Getter
     @Setter
     public static class CreateReview {
-        Integer score;
+        Float score;
         String content;
         Long userId;
         Long restaurantId;
